requirements.txt
streamlit
langchain
langchain-google-genai
faiss-cpu
python-dotenv
